bgl-python/build/Jamfile:   - Link against Boost.Python   - Install _support into the right place
bgl-python/build/distributed/Jamfile:
  - Pull in bsp_process_group implementation
  - Link against Boost.Python

bgl-python/python/boost/graph/distributed/__init__.py,
bgl-python/python/boost/graph/__init__.py,
bgl-python/python/boost/__init__.py:
  - Make these directories into Python modules

bgl-python/boost/graph/python/distributed/boman_et_al_graph_coloring.hpp,
bgl-python/src/distributed/boman_et_al_graph_coloring.cpp:
  - Use "int" instead of "std::size_t". The madness need not extent into
    Python

bgl-python/boost/graph/python/distributed/dijkstra_shortest_paths.hpp,
bgl-python/boost/graph/python/distributed/dijkstra_shortest_paths.hpp,
bgl-python/src/distributed/dijkstra_shortest_paths.cpp:
  - Use "float" instead of "double"

bgl-python/boost/graph/python/connected_components.hpp,
bgl-python/src/distributed/connected_components.cpp:
  - Use "int" instead of "std::size_t"

bgl-python/src/distributed/Graph.cpp:
  - Expose vertex->color distributed property map

+++ src/distributed/dijkstra_shortest_paths.cpp
@@ -23,7 +23,7 @@ void export_dijkstra_shortest_paths()
   typedef property_map<Graph, vertex_index_t>::type VertexIndexMap;
   typedef vector_property_map<vertex_descriptor, VertexIndexMap>
     VertexPredecessorMap;
-  typedef vector_property_map<double, VertexIndexMap>
+  typedef vector_property_map<float, VertexIndexMap>
     VertexDistanceMap;
   typedef vector_property_map<default_color_type, VertexIndexMap>
     VertexColorMap;
